create ScrimSectionHeader
